BossStage <-> Player 간 파티클 오류 수정

+++ Source/UE4_BeatNight/Public/Enemy.h
@@ -53,10 +53,10 @@ protected:
 	
 	/** 최대 체력 */
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Enemy|Props", meta=(AllowPrivateAccess=true))
-	int32 MaxHealth;
+	float MaxHealth;
 	/** 체력 */
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Enemy|Props", meta=(AllowPrivateAccess=true))
-	int32 Health;
+	float Health;
 	/** 데미지 */
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Enemy|Props", meta=(AllowPrivateAccess=true))
 	float EnemyDamage;
